+++ layout/secure2r.php
@@ -29,6 +29,36 @@
 $OUTPUT->optional_jquery();
 // Get the HTML for the settings bits.
 $html = theme_campus_get_html_for_settings($OUTPUT, $PAGE);
+
+$rtl = right_to_left();  // To know if to add 'pull-right' and 'desktop-first-column' classes in the layout for LTR.
+$hassidepre = $PAGE->blocks->is_known_region('side-pre');
+$hassidepost = $PAGE->blocks->is_known_region('side-post');
+if ($hassidepre) {
+    $useblock = 'side-pre';
+    /*
+     This deals with the side to show the blocks on.
+     If we have a 'side-pre' then the blocks are on the right for LTR and left for RTL.
+    */
+    if ($rtl) {
+        $right = false;
+    } else {
+        $right = true;
+    }
+} else if ($hassidepost) {
+    $useblock = 'side-post';
+    /*
+     This deals with the side to show the blocks on.
+     If we have a 'side-post' then the blocks are on the left for LTR and right for RTL.
+    */
+    if ($rtl) {
+        $right = true;
+    } else {
+        $right = false;
+    }
+} else {
+    $useblock = false;
+}
+
 echo $OUTPUT->doctype() ?>
 <html <?php echo $OUTPUT->htmlattributes(); ?>>
 <head>
@@ -40,7 +70,7 @@ echo $OUTPUT->doctype() ?>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 </head>
 
-<body <?php echo $OUTPUT->body_attributes(); ?>>
+<body <?php echo $OUTPUT->body_attributes('two-column'); ?>>
 
 <?php
 echo $OUTPUT->standard_top_of_body_html();
@@ -54,15 +84,22 @@ require_once(dirname(__FILE__).'/tiles/'.$OUTPUT->get_header_file());
     </header>
 
     <div id="page-content" class="row-fluid">
-        <div id="region-bs-main-and-pre" class="span9">
-            <div class="row-fluid">
-                <section id="region-main" class="span8 pull-right">
-                    <?php echo $OUTPUT->main_content(); ?>
-                </section>
-                <?php echo $OUTPUT->blocks('side-pre', 'span4 desktop-first-column'); ?>
-            </div>
-        </div>
-        <?php echo $OUTPUT->blocks('side-post', 'span3'); ?>
+        <?php if ($useblock) { ?>
+        <section id="region-main" class="span9<?php if (!$right) { echo ' pull-right'; } ?>">
+        <?php } else { ?>
+        <section id="region-main" class="span12">
+        <?php }
+            echo $OUTPUT->main_content(); ?>
+        </section>
+        <?php
+        if ($useblock) {
+            $classextra = '';
+            if (!$right) {
+                $classextra = ' desktop-first-column';
+            }
+            echo $OUTPUT->campussingleblocks('span3'.$classextra);
+        }
+        ?>
     </div>
 
     <?php require_once(dirname(__FILE__).'/tiles/footer.php'); ?>